+++ packages/nimble-components/src/theme-provider/tests/tokens.stories.ts
@@ -1,11 +1,13 @@
 import type { Meta, StoryObj } from '@storybook/html';
-import { html, repeat, when } from '@microsoft/fast-element';
+import { html, repeat, ViewTemplate, when } from '@microsoft/fast-element';
 import { PropertyFormat } from './types';
 import { createUserSelectedThemeStory } from '../../utilities/tests/storybook';
 import {
     tokenNames,
     cssPropertyFromTokenName,
-    scssPropertyFromTokenName
+    scssPropertyFromTokenName,
+    TokenSuffix,
+    suffixFromTokenName
 } from '../design-token-names';
 import { comments } from '../design-token-comments';
 import '../../all-components';
@@ -29,7 +31,7 @@ interface TokenArgs {
 }
 
 const metadata: Meta = {
-    title: 'Tokens/Property Names',
+    title: 'Tokens/Theme-aware Tokens',
     parameters: {
         docs: {
             description: {
@@ -41,8 +43,89 @@ const metadata: Meta = {
 
 export default metadata;
 
+const computedCSSValueFromTokenName = (tokenName: string): string => {
+    return getComputedStyle(document.documentElement).getPropertyValue(
+        cssPropertyFromTokenName(tokenName)
+    );
+};
+
+const colorTemplate = html<TokenName>`
+    <div
+        title="${x => computedCSSValueFromTokenName(tokenNames[x])}"
+        style="
+        display: inline-block;
+        height: 24px;
+        width: 24px;
+        border: 1px solid black;
+        background-color: var(${x => cssPropertyFromTokenName(tokenNames[x])});
+    "
+    ></div>
+`;
+
+const rgbColorTemplate = html<TokenName>`
+    <div
+        title="${x => computedCSSValueFromTokenName(tokenNames[x])}"
+        style="
+        display: inline-block;
+        height: 24px;
+        width: 24px;
+        border: 1px solid black;
+        background-color: rgba(var(${x => cssPropertyFromTokenName(tokenNames[x])}), 1.0);
+    "
+    ></div>
+`;
+
+const stringValueTemplate = html<TokenName>`
+    <div style="display: inline-block;">
+        ${x => computedCSSValueFromTokenName(tokenNames[x])}
+    </div>
+`;
+
+const fontTemplate = html<TokenName>`
+    <div
+        style="
+        display: inline-block;
+        font: var(${x => cssPropertyFromTokenName(tokenNames[x])});
+    "
+    >
+        Nimble
+    </div>
+`;
+
+/* eslint-disable @typescript-eslint/naming-convention */
+const tokenTemplates: {
+    readonly [key in TokenSuffix]: ViewTemplate<TokenName>;
+} = {
+    Color: colorTemplate,
+    RgbPartialColor: rgbColorTemplate,
+    FontColor: colorTemplate,
+    FontLineHeight: stringValueTemplate,
+    FontWeight: stringValueTemplate,
+    FontSize: stringValueTemplate,
+    TextTransform: stringValueTemplate,
+    FontFamily: stringValueTemplate,
+    Font: fontTemplate,
+    Size: stringValueTemplate,
+    Width: stringValueTemplate,
+    Height: stringValueTemplate,
+    Delay: stringValueTemplate,
+    Padding: stringValueTemplate
+};
+/* eslint-enable @typescript-eslint/naming-convention */
+
+const templateForTokenName = (
+    tokenName: TokenName
+): ViewTemplate<TokenName> => {
+    const suffix = suffixFromTokenName(tokenName);
+    if (suffix === undefined) {
+        throw new Error(`Cannot identify suffix for token: ${tokenName}`);
+    }
+    const template = tokenTemplates[suffix];
+    return template;
+};
+
 // prettier-ignore
-export const propertyNames: StoryObj<TokenArgs> = {
+export const themeAwareTokens: StoryObj<TokenArgs> = {
     parameters: {
         controls: { hideNoControlsWarning: true }
     },
@@ -67,12 +150,16 @@ export const propertyNames: StoryObj<TokenArgs> = {
                 color: var(${groupHeaderFontColor.cssCustomProperty});
                 text-transform: var(${groupHeaderTextTransform.cssCustomProperty});
             }
-            td { padding: 10px;}
+            td { 
+                padding: 10px;
+                height: 32px;
+            }
         </style>
         <table>
             <thead>
                 <tr>
                     <th>${x => x.propertyFormat} Property</th>
+                    <th>Preview</th>
                     <th>Description</th>
                 </tr>
             </thead>
@@ -87,6 +174,7 @@ export const propertyNames: StoryObj<TokenArgs> = {
                             ${x => scssPropertyFromTokenName(tokenNames[x])}
                         `)}
                     </td>
+                    <td>${x => templateForTokenName(x)}</td>
                     <td>${x => comments[x]}</td>
                 </tr>
             `)}
@@ -94,3 +182,5 @@ export const propertyNames: StoryObj<TokenArgs> = {
         </table>
     `)
 };
+
+themeAwareTokens.storyName = 'Theme-aware Tokens';